# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sqlx-todos",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/stories/{story_id}/files": {
      "get": {
        "tags": [
          "File"
        ],
        "summary": "List files for a story.",
        "operationId": "get_files",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The parent story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A file metadata array for the story",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoryFile"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The parent story was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "File"
        ],
        "summary": "Add files to a story.",
        "operationId": "add_files",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The parent story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A metadata array for the uploaded files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoryFile"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The parent story was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
    "/stories/{story_id}/files/{file_id}": {
      "get": {
        "tags": [
          "File"
        ],
        "summary": "Get file metadata.",
        "operationId": "get_file",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The parent story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "description": "The id of the file metadata",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryFile"
                }
              }
            }
          },
          "404": {
            "description": "The file was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "File"
        ],
        "summary": "Delete a file",
        "operationId": "delete_file",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The parent story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "description": "The id of the file to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The file was deleted successfully"
          },
          "404": {
            "description": "The file was not found"
          }
        }
      }
    },
    "/stories/{story_id}/files/{file_id}/contents": {
      "get": {
        "tags": [
          "File"
        ],
        "summary": "Download file contents.",
        "operationId": "download_file",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The parent story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "description": "The id of the file to download",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The contents of the file"
          },
          "404": {
            "description": "The file was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Errors": {
        "type": "object",
        "description": "The type sent as an error response to the client.",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "FileUpload": {
        "type": "object",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        }
      },
...
      "StoryFile": {
        "type": "object",
        "required": [
          "id",
          "story_id",
          "name",
          "size",
          "content_type",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "story_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}